power: supply: Add driver for Qualcomm SMBCharger
Add a driver for Qualcom Switch-Mode Battery Charger, found on
PMICs such as PMI8994. The SMBCharger supports charging from
USB or DC input, as well as generating VBUS for USB OTG with
a boost regulator. It can also be operated along with a parallel
charger for improved high voltage/current fast charging performance.

At the moment, this driver only includes support for operating
the OTG regulator as well as partial support for the USB charge path.
This driver also provides extcon functionality, reporting states of
USB and USB_HOST with VBUS, as well as charge port types.

Co-developed-by: Alejandro Tafalla <atafalla@dnyon.com>
Signed-off-by: Yassine Oudjana <y.oudjana@protonmail.com>