arm64: dts: qcom: msm8996-xiaomi-scorpio: Use MSM8996 Pro
The Xiaomi Mi Note 2 has the MSM8996 Pro SoC. Rename the dts
to match, include msm8996pro.dtsi, and add the qcom,msm8996pro
compatible. To do that, the msm8996.dtsi include in msm8996-xiaomi-common
has to be moved to msm8996-xiaomi-gemini, the only device that needs it
included after this change. The msm-id is also removed as it is now defined
in msm8996pro.dtsi.

Signed-off-by: Yassine Oudjana <y.oudjana@protonmail.com>